reorganizing SearchForm Component
