+++ src/Departures.jsx
@@ -9,64 +9,59 @@ class UnconnectedDepartures extends Component {
     let lng = this.props.language;
     return (
       <div className="global-container">
-        <div className="search-bus-container">
-          <PaginationButtons />
-        </div>
+        <PaginationButtons />
         <div>
-          {this.props.departures !== undefined &&
-          this.props.operators !== undefined
-            ? this.props.departures
-                .slice(this.props.page * 3, this.props.page * 3 + 3)
-                .map(bus => {
-                  return (
-                    <div className="departure-result-container">
-                      {this.props.operators.map(op => {
-                        if (op.id === bus.operator_id) {
-                          return (
-                            <div>
-                              <div className="operator-price">
-                                <b>{op.display_name.toUpperCase()}</b>
-                                <br />
-                                <div className="label-bus-results" id="price">
-                                  {bus.prices.total}
-                                  {" $" + bus.prices.currency.slice(0, 2)}
-                                </div>{" "}
-                              </div>
-                              <div>
-                                <div className="label-bus-results">
-                                  {lng === "Fr"
-                                    ? t("Departure time:")
-                                    : "Departure time:"}{" "}
-                                </div>
-                                <div className="departure-arrival-result">
-                                  {bus.departure_time
-                                    .split("T")
-                                    .join(" @ ")
-                                    .slice(0, -3)}
-                                  {" from " + this.props.origin}
-                                </div>
-                              </div>
-                              <div>
-                                <div className="label-bus-results">
-                                  {lng === "Fr"
-                                    ? t("Arrival time:")
-                                    : "Arrival time:"}{" "}
-                                </div>
-                                <div className="departure-arrival-result">
-                                  {bus.arrival_time
-                                    .split("T")
-                                    .join(" @ ")
-                                    .slice(0, -3)}
-                                </div>
-                              </div>
+          {this.props.departures
+            .slice(this.props.page * 3, this.props.page * 3 + 3)
+            .map(bus => {
+              return (
+                <div className="departure-result-container">
+                  {this.props.operators.map(op => {
+                    if (op.id === bus.operator_id) {
+                      return (
+                        <div>
+                          <div className="operator-price">
+                            <b>{op.display_name.toUpperCase()}</b>
+                            <br />
+                            <div className="label-bus-results" id="price">
+                              {bus.prices.total}
+                              {" $" + bus.prices.currency.slice(0, 2)}
+                            </div>{" "}
+                          </div>
+                          <div>
+                            <div className="label-bus-results">
+                              {lng === "Fr"
+                                ? t("Departure time:")
+                                : "Departure time:"}{" "}
                             </div>
-                          );
-                        }
-                      })}
-                    </div>
-                  );
-                })
-            : null}
+                            <div className="departure-arrival-result">
+                              {bus.departure_time
+                                .split("T")
+                                .join(" @ ")
+                                .slice(0, -3)}
+                              {" from " + this.props.origin}
+                            </div>
+                          </div>
+                          <div>
+                            <div className="label-bus-results">
+                              {lng === "Fr"
+                                ? t("Arrival time:")
+                                : "Arrival time:"}{" "}
+                            </div>
+                            <div className="departure-arrival-result">
+                              {bus.arrival_time
+                                .split("T")
+                                .join(" @ ")
+                                .slice(0, -3)}
+                            </div>
+                          </div>
+                        </div>
+                      );
+                    }
+                  })}
+                </div>
+              );
+            })}
         </div>
       </div>
     );
